chore: cleanup dashboard page setup (#10652)
